Extract sample to separate submodule (#104)
